spongecake-sdk/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spongecake"
version = "0.1.13"
description = "Spongecake is the easiest way to launch OpenAI computer use agents."
authors = [
  { name = "Terrell Marshall", email = "terrell@passage-team.com" },
  { name = "Aditya Nadkarni", email = "aditya@passage-team.com" }
]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
dependencies = [
  "requests>=2.0.0",          # keep if the rest of your SDK still uses it
  "httpx>=0.27.0",            # ← NEW
  "docker>=6.0.0",
  "openai>=1.66.3",
  "python-dotenv>=0.17.1",
  "marionette-driver>=3.0.0",
  "pyautogui>=0.9.54",
  "posthog>=3.0.0",           # For telemetry service
]

# Optional: Additional project metadata
keywords = ["sdk", "browser", "fastapi"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[tool.setuptools.packages.find]
where = ["."]
include = [".", "spongecake*"]
